Reducing the number of changes in the program
